+++ monstertruck-meshing/src/filters/optimizing/mod.rs
@@ -453,18 +453,6 @@ impl IntVector for [i64; 3] {
         ]
     }
 }
+
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn into_vertices(iter: &[usize]) -> Vec<Vertex> { iter.iter().map(|i| i.into()).collect() }
-
-    #[test]
-    fn degenerate_polygon_test() {
-        let poly = into_vertices(&[0, 1, 2, 0, 3, 4, 5, 6, 3, 7, 8, 9]);
-        let polys = split_into_nondegenerate(poly);
-        assert_eq!(polys[0], into_vertices(&[0, 1, 2]));
-        assert_eq!(polys[1], into_vertices(&[3, 4, 5, 6]));
-        assert_eq!(polys[2], into_vertices(&[3, 7, 8, 9, 0]));
-    }
-}
+mod tests;
